+++ iron-core/src/main/java/io/axway/iron/core/spi/testing/TransientSnapshotStore.java
@@ -1,6 +1,7 @@
 package io.axway.iron.core.spi.testing;
 
 import java.io.*;
+import java.math.BigInteger;
 import java.util.*;
 import com.google.common.collect.ImmutableList;
 import io.axway.iron.spi.storage.SnapshotStore;
@@ -8,7 +9,7 @@ import io.axway.iron.spi.storage.SnapshotStore;
 class TransientSnapshotStore implements SnapshotStore {
 
     @Override
-    public OutputStream createSnapshotWriter(long transactionId) throws IOException {
+    public OutputStream createSnapshotWriter(BigInteger transactionId) throws IOException {
         return new OutputStream() {
             @Override
             public void write(int b) throws IOException {
@@ -18,17 +19,17 @@ class TransientSnapshotStore implements SnapshotStore {
     }
 
     @Override
-    public InputStream createSnapshotReader(long transactionId) throws IOException {
+    public InputStream createSnapshotReader(BigInteger transactionId) throws IOException {
         throw new IOException("Snapshot for transaction id=" + transactionId + " has not been found");
     }
 
     @Override
-    public List<Long> listSnapshots() {
+    public List<BigInteger> listSnapshots() {
         return ImmutableList.of();
     }
 
     @Override
-    public void deleteSnapshot(long transactionId) {
+    public void deleteSnapshot(BigInteger transactionId) {
         // nothing to do
     }
 }
